Add a WirelessNetworkItem class which represents a wireless network for which no Connection exists. Introduce an AbstractConnectableItem to the item hierarchy:
    AbstractConnectableItem
       ^                ^
       |                |
ConnectionItem   WirelessNetworkItem
       ^
       |
WirelessConnectionItem

The plan is not to show these in the popup unless there are no
Connections which could be activated, but at present the 3 strongest
networks are displayed, to stimulate a fix for the horrible layout
problems that adding and removing these items from the layout
at runtime causes.

There is no support in the service for handling clicks on these items
yet, either.

CCMAIL: sebas@kde.org
CCMAIL: cblauvelt@gmail.com