Rework the painting of the applet a bit
we'll simply route to a method per device so we can do custom painting
of them individually and per state.

The sorting of the interfaces (and thus the decision which one should be
painted seems to be broken). While it's fine when I start the applet in
plasmoidviewer, disabling and enabling the wireless leaves me with the
wired icon. WHen enabling, the correct iface is used and hence the
wireless interface painted as not connected. As soon as it's connected,
it's going back to the wired item.

CCMAIL:wstephenson@kde.org