backport r883506: icons, and other ui work
We're now compressing the notficiations when a new wireless network
appears by (re-)starting a timer when a new network appears and
collecting the network's essids in the meantime. This prevents the event
notification storm when you start the applet and there are existing
networks found. Much, much better this way ... :)

Also, fix lots of icons. media-playback* is semantically incorrect, use
network-(dis)connect instead.