better alignment, less empty updates to wifigroup
The icons and meters of  the interfaceitem now align nicely with the
list of connections. wirelessnetworkitems now look well-aligned,
connectionitems not yet (but quite a bit better).

There were a couple of updateNetworks() happening when the interface is
initialized. Now at least on startup, things seem to work (i.e. I'm
getting a list of networks to connect to, and can click on my
preconfig'ed one, and huzzah ... it connects. When connected, the list
of wireless connections is empty (so in theory, I cannot switch
networks). Need to look into that next.

And add some excessive debugging to see what's going on.