Knock out any 802.1x settings found when saving WPA-PSK security.
These settings were being written automatically because the
KConfigDialogManager contains all widgets for all security types. 

In the medium term we need to do this cleanly instead of allowing excess
configurations like this to be written then having to trim them.

This may also need doing for different flavours of WPA-EAP.

Helmut, is this what we decided on Friday?
CCMAIL: hschaa@suse.de