Save knetworkmanagerrc with the same contents as the setting's own file.
Add a parameter to ConnectionWidget that supplies a default connection
name.  It is necessary to set special default values in a connection in
the widget's readConfig() method, as ConnectionPreferences::load()
performs the file read and UI initialisation in one go.  Therefore,
preset the special value in the widget's ctor, and apply it in
readConfig() after the defaults from the config file have been loaded.

Allow a presupplied SSID as above.

Stop marshalling wireless parameters that are not in the UI to DBus at
all.

Add a new 'Automatic' special value for MTU==0 in the UI.  Default
Wireless MTU to 0 which is left out of the settings map when marshalled
for automatic MTU setting.

Remove Wired and WirelessPreferences save() and load() reimpls, they
were doing nothing.