updateGeometry() works much better for layout size changes than invalidating the layout.
This fixes one part of the sizing problems, the extenderitem's size is
updated when you move the extender around. Next step would be to make
the extenderapplet update its geometry after the extenderitem's size
changes. Is there a signal I should use, or what is the best way to do
this?