Add some debugging output
The retrieval of m_activeConnections is broken on my machine it seems,
dunno if it works for anybody else. Given that the bugreports suggest it
works for others, and that they see bugs I'm not even able to reproduce
(since m_activeConnections is empty), I suspect a local setup issue.
I've no clue how to sort this out, the config file in the /etc/dbus dir
is the same we ship. This kind of blocks me from fixing other issues.