Check in the first classes of a library for making backend independent network management services.
NetworkSettings was moved and renamed to ConnectionList.
KConfigToService was moved and renamed to ConnectionListPersistence.

QtTest unit tests for those classes which are statically testable are in
libs/service/tests.  The NetworkInterface* classes depend on the
system's networking as reported by Solid::Control which prevents testing
while Solid::Control does not have a scriptable fake backend.