Add API to the Activatables to provide wireless security data and signal strength.
This is achieved by putting this interface (and moving SSID) in a new base class,
WirelessItem, and letting WirelessNetworkItem and
WirelessInterfaceConnection inherit from it.

Update the WirelessNetworkInterfaceActivatableProvider to obtain this
information from Solid and hook up signals for updates.

Update ActivatableDebug to have a basic way to see signal strength
changes.