start to bring back config
Whether to show a certain type of connection is now controllable,
it needs a bit more work (especially the crash after removing
a tab), and isn't done for anything else but wireless.

This will also make it easier to make showing and hiding of
certain interfaceitems easier.

Also nuke dbus connection breakdown