Make this guy a Plasma::Label,
It shouldn't be separately clickable, as this brings some interaction problems
with it. Now the whole area of this widget is clickable and triggers the details
view.

We should probably expose the setPixmap(QPixmap) method in Plasma::Label's API,
as this functionality is very useful if you want to display a custom-painted
pixmap we don't have the path of.

CCMAIL:plasma-devel@kde.org