Active connections now are bold
The default route gets a smilie. Initially, we planned to make the default
route bold, but I think this is much nicer, especially when you have multiple
connections. When there's only one connection, the active one is the default
one which is bold, so it's not a big deal for the common case.