Enable Plasmoid by default, don't autostart KNM if not asked for
This means we now basically switch the installation to the Plasmoid,
which will (should?!) now work out of the box on new installs, no
problems with the DBus module not being loaded anymore.

We're installing the knm autostart file conditionally now (cmake
-DINSTALL_KNM_AUTOSTART=ON) and load the NM kded module by default.

Let's see how this fares. As discussed with wstephenson

CCMAIL:wstephenson@kde.org