Fix tests build with Qt 4.7
Changes introduced by
http://qt.gitorious.org/qt/qt/commit/f1d7cbd9efbdade9df5496dc684c2f7d22db7378
modifies the macro QDBUS_EXPORT to Q_DBUS_EXPORT. This commit updates the code
where the macro is used and also adds an ifdef to keep it compiling with
Qt < 4.7.

I know the default scenario is to compile with Qt 4.6, but as 4.7 is about
to be released, I chose to adjust the code according to the upcoming version.