Try harder to find a unique UUID for new connnections
Fixes bug where the new connection editor would be opened with the
values of an existing connnection.

QUuid::createUuid() cannot be guaranteed to be unique, and in fact it
often produces an existing value the first time. This leads to the new
Connection to be initialized with a Uuid that it already used for
another stored connection. The new Connection then reads this config and
overwrites the values passed as arguments.

We now check the newly created Uuid is already existing on the system,
and if it is, we try to create a new one. We won't try forever though,
since hanging the system is not really an option. The odds that this
happens are really low, anyway.

This bug has been annoying me and many others, so glad it's fixed. It
took a while to figure out what exactly is going wrong.