Hint at the reason for disconnect
After half an hour of debugging if there's anything wrong with
networkmanager, or the plasmoid, I've found out that wifi doesn't work
because the hardware switch is disabled. PEBKAC top-to-bottom.

Now we show the reason why the network doesn't work in the tooltip.
(There's really nothing we can do automatically when hardware is
disabled.)