vpn: fix UI plugin loading for systemwide connections
KNM was using pluginName to lookup settings UI plugin. For user
connection pluginName was stored in ini file, but for systemwide
connections it is not available. So use service-type vpn property
which is always present.

Modify plugin service definition to enable lookup on
X-NetworkManager-Services property and lookup it using service-type
from connection. Set property to full name in all plugins to
make lookup easier.

Signed-off-by: Andrey Borzenkov <arvidjaar@gmail.com>