Revert "Fix delayed DBus reply logic (I guess), but GetSecrets still does not"
This reverts commit e428bf150989269ba0a7f9ab050cc76d6e1f6f2c.

The whole SecretStorage/SecretsProvider class works asynchronously, so this breaks
more than it fixes (editing of connections for example, if you edit the connection
once the secrets are saved, on second edit they are retrieved - successfully, tested
with plain text storage). Besides, delayed reply logic was working in
backends/NetworkManager/busconnection.cpp, so it is highly unlikely that this is the
problem. Also, here the return value is the argument of the method, I don't think
that's what we want.