Get secrets at first try when editing.
The problem was that NMDBusSettingsConnectionProvider maintains two list
of connections, one passed to the constructor and another created
in a private class (d class). When
NMDBusSettingsConnectionProvider::onConnectionSecretsArrived gets the
secrets it was updating only the connection in the d class class but
manageconnectionwidget uses the other connection list.

To make things simpler I added the secrets to the signal
manageconnectionwidget already listen to.

Maybe in the future we could use only one list of connections instead of
two.