Add version to kcm. The version is the branch name (if not master) plus the date when the last commit was done, for example: nm09-20110619. This will make it easier to know which version users are using.
Problems are:

1. Every developer must add the script hook that updates the
plasma_nm_version.h file. Is there are way to add it to the repository
server to that everybody downloads it automatically?

2. The current script hook updates the files after the commit has been
created :-/ So you will have to do a second commit to send the updated
plasma_nm_version.h to the repository.

I am using .git/hooks/pre-commit because I do not know any other that is
run during a commit, unfortunately git run it only after it has already
scanned source code files.