Send an error DBus reply to NM if secrets retrieval fails
Currently, we just send an empty settings map. However, at least
for VPN connections, this doesn't cause tem to fail, but they are
stuck in activating state, which happens for example if the user
cancels the auth dialog. Sending an error reply to NM will cause
the connection to fail immediately.