rework VPN status overlay icon
we need to trigger a repaint on activation/deactivation of a VPN connection,
otherwise, when using a wired connection, which does not receive
signalStrengthChanged signals and thus does not repaint every few seconds,
the VPN overlay will only be shown after a click on the NetworkManagement icon.

I guess this commit also improves performance, as we don't need to iterate
over all available VPN connections at every repaint, but only over the active
connections, and even that only when the tooltip is about to be shown.

This commit also makes an ancient hack, which originated due to not working
autorelaying of signals on our session bus service, a little less hacky

fixes a crash in plasma when kded is restarted