Fix creation of Bluetooth PAN connections.
NM refuses to add Bluetooth PAN connections if they include settings for
Gsm, Ppp and Serial configurations. On the other hand those settings are
required by Bluetooth DUN connections. Because of the way our internal
libraries are coded it is not that trivial to check the correct type
(DUN / PAN) before creating a connection, so I am removing the unneeded
settings when creating / editing a connection.

BUG: 285637